requirements.txt
Django==4.1.4
djangorestframework==3.14.0
gTTS==2.3.0
gunicorn==20.1.0
num2words==0.5.12
whitenoise==6.2.0

